refactor(test): move project management calls to server
