+++ tests/api/main.rs
@@ -1,6 +1,5 @@
 mod manage_parts;
 mod register_part;
 mod register_project;
-mod rest;
 mod server;
 mod server_health;
